Move Cart and Timer to cartcontext, and logic to ContextWrapper
