+++ src/Components/context/cartcontext.js
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+
+export const CartContext = createContext({
+    cart: [],
+    setCart: () => {},
+
+    cartTimer: 0,
+    setCartTimer: () => {},
+
+    timerRunning: false,
+    setTimerRunning: () => {},
+});
