+++ src/Pages/Seat.js
@@ -6,10 +6,11 @@ import { Link } from 'react-router-dom';
 import { useEffect, useContext } from 'react';
 import { useHistory } from 'react-router-dom';
 import { DataContext } from '../Components/context/datacontext';
+import { CartContext } from '../Components/context/cartcontext';
 
-function Seat({seats, setSeats, cart, setCart, isLoading, setCartTimer, timerRunning, setTimerRunning}) {
+function Seat({seats, setSeats, isLoading}) {
 const {movies, theaters, movieID, theaterID} = useContext(DataContext)
-
+const {cart, setCart, timerRunning, setTimerRunning, setCartTimer} = useContext(CartContext);
 
 const history = useHistory();
 useEffect(() => {
